Finish converting adventure card and adventure spell
