+++ Mage.Common/src/main/java/mage/view/CardView.java
@@ -432,21 +432,21 @@ public class CardView extends SimpleCardView {
                 fullCardName = mainCard.getLeftHalfCard().getName() + MockCard.MODAL_DOUBLE_FACES_NAME_SEPARATOR + mainCard.getRightHalfCard().getName();
                 this.manaCostLeftStr = mainCard.getLeftHalfCard().getManaCostSymbols();
                 this.manaCostRightStr = mainCard.getRightHalfCard().getManaCostSymbols();
-            } else if (card instanceof AdventureCard) {
+            } else if (card instanceof SingleFaceSplitCard) {
                 this.isSplitCard = true;
-                AdventureCard adventureCard = ((AdventureCard) card);
-                leftSplitName = adventureCard.getName();
-                leftSplitCostsStr = String.join("", adventureCard.getManaCostSymbols());
-                leftSplitRules = adventureCard.getSharedRules(game);
-                leftSplitTypeLine = getCardTypeLine(game, adventureCard);
-                AdventureCardSpell adventureCardSpell = adventureCard.getSpellCard();
-                rightSplitName = adventureCardSpell.getName();
-                rightSplitCostsStr = String.join("", adventureCardSpell.getManaCostSymbols());
-                rightSplitRules = adventureCardSpell.getRules(game);
-                rightSplitTypeLine = getCardTypeLine(game, adventureCardSpell);
-                fullCardName = adventureCard.getName() + MockCard.ADVENTURE_NAME_SEPARATOR + adventureCardSpell.getName();
-                this.manaCostLeftStr = adventureCard.getManaCostSymbols();
-                this.manaCostRightStr = adventureCardSpell.getManaCostSymbols();
+                SingleFaceSplitCard mainCard = ((SingleFaceSplitCard) card);
+                leftSplitName = mainCard.getName();
+                leftSplitCostsStr = String.join("", mainCard.getManaCostSymbols());
+                leftSplitRules = mainCard.getSharedRules(game);
+                leftSplitTypeLine = getCardTypeLine(game, mainCard);
+                SingleFaceSplitCardSpell splitCardSpell = mainCard.getSpellCard();
+                rightSplitName = splitCardSpell.getName();
+                rightSplitCostsStr = String.join("", splitCardSpell.getManaCostSymbols());
+                rightSplitRules = splitCardSpell.getRules(game);
+                rightSplitTypeLine = getCardTypeLine(game, splitCardSpell);
+                fullCardName = mainCard.getName() + MockCard.SINGLE_FACE_SPLIT_NAME_SEPARATOR + splitCardSpell.getName();
+                this.manaCostLeftStr = mainCard.getManaCostSymbols();
+                this.manaCostRightStr = splitCardSpell.getManaCostSymbols();
             } else if (card instanceof MockCard) {
                 // deck editor cards
                 fullCardName = ((MockCard) card).getFullName(true);
