+++ Mage/src/main/java/mage/cards/mock/MockCard.java
@@ -20,7 +20,7 @@ import java.util.List;
  */
 public class MockCard extends CardImpl implements MockableCard {
 
-    public static String ADVENTURE_NAME_SEPARATOR = " // ";
+    public static String SINGLE_FACE_SPLIT_NAME_SEPARATOR = " // ";
     public static String MODAL_DOUBLE_FACES_NAME_SEPARATOR = " // ";
 
     // Needs to be here, as it is normally calculated from the
@@ -156,7 +156,7 @@ public class MockCard extends CardImpl implements MockableCard {
         }
 
         if (adventureSpellName != null) {
-            return getName() + ADVENTURE_NAME_SEPARATOR + adventureSpellName;
+            return getName() + SINGLE_FACE_SPLIT_NAME_SEPARATOR + adventureSpellName;
         } else if (isModalDoubleFacedCard) {
             return getName() + MODAL_DOUBLE_FACES_NAME_SEPARATOR + this.getSecondCardFace().getName();
         } else {
