+++ Mage/src/main/java/mage/game/GameState.java
@@ -1639,14 +1639,14 @@ public class GameState implements Serializable, Copyable<GameState> {
             copiedParts.add(rightCopied);
             // sync parts
             ((ModalDoubleFacedCard) copiedCard).setParts(leftCopied, rightCopied);
-        } else if (copiedCard instanceof AdventureCard) {
+        } else if (copiedCard instanceof SingleFaceSplitCard) {
             // right
-            AdventureCardSpell rightOriginal = ((AdventureCard) copiedCard).getSpellCard();
-            AdventureCardSpell rightCopied = rightOriginal.copy();
+            SingleFaceSplitCardSpell rightOriginal = ((AdventureCard) copiedCard).getSpellCard();
+            SingleFaceSplitCardSpell rightCopied = rightOriginal.copy();
             prepareCardForCopy(rightOriginal, rightCopied, newController);
             copiedParts.add(rightCopied);
             // sync parts
-            ((AdventureCard) copiedCard).setParts(rightCopied);
+            ((SingleFaceSplitCard) copiedCard).setParts(rightCopied);
         }
 
         // main part prepare (must be called after other parts cause it change ids for all)
